extract db set up to separate class
